Refactor typedstream module organization
